+++ sources/Google.Solutions.Ssh/Native/UnsafeNativeMethods.cs
@@ -113,6 +113,21 @@ namespace Google.Solutions.Ssh.Native
     [StructLayout(LayoutKind.Sequential)]
     public struct LIBSSH2_STAT
     {
+        //
+        // dt libssh2_struct_stat
+        //    +0x000 st_dev           : Uint4B
+        //    +0x004 st_ino           : Uint2B
+        //    +0x006 st_mode          : Uint2B
+        //    +0x008 st_nlink         : Int2B
+        //    +0x00a st_uid           : Int2B
+        //    +0x00c st_gid           : Int2B
+        //    +0x010 st_rdev          : Uint4B
+        //    +0x018 st_size          : Int8B
+        //    +0x020 st_atime         : Int8B
+        //    +0x028 st_mtime         : Int8B
+        //    +0x030 st_ctime         : Int8B
+        //
+
         public uint st_dev;
         public ushort st_ino;
         public ushort st_mode;
@@ -120,10 +135,10 @@ namespace Google.Solutions.Ssh.Native
         public short st_uid;
         public short st_gid;
         public uint st_rdev;
-        public uint st_size;
-        public uint st_atime;
-        public uint st_mtime;
-        public uint st_ctime;
+        public ulong st_size;
+        public long st_atime;
+        public long st_mtime;
+        public long st_ctime;
     }
 
     internal static class UnsafeNativeMethods
